requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.10.0
gradio==5.9.0
gradio_client==1.5.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.2.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
paypalrestsdk==1.13.3
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.3
safehttpx==0.1.6
safetensors==0.4.5
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uuid==1.30
uvicorn==0.34.0
websockets==14.1
